requirements.txt
asttokens==2.0.8
azure-functions==1.12.0
backcall==0.2.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.1
ipykernel==6.16.1
ipython==8.5.0
jedi==0.18.1
joblib==1.2.0
jupyter_client==7.4.3
jupyter_core==4.11.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
pyzmq==24.0.1
scikit-learn==0.24.2
scipy==1.9.3
six==1.16.0
stack-data==0.5.1
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
wcwidth==0.2.5
